refactor: rename custommaps package to custom_maps (match sampling_design convention)
