+++ backend/src/main.py
@@ -17,7 +17,7 @@ from src.annotation.router import router as annotations_router
 from src.auth.router import router as auth_router
 from src.campaigns.router import router as campaigns_router
 from src.config import get_settings
-from src.custommaps.router import router as custom_maps_router
+from src.custom_maps.router import router as custom_maps_router
 from src.database import SessionLocal
 from src.imagery.proxy_router import router as imagery_proxy_router
 from src.imagery.router import router as imagery_router
