+++ backend/src/models.py
@@ -5,4 +5,4 @@ from src.campaigns.models import *  # noqa
 from src.annotation.models import *  # noqa
 from src.imagery.models import *  # noqa
 from src.timeseries.models import *  # noqa
-from src.custommaps.models import *  # noqa
+from src.custom_maps.models import *  # noqa
