+++ backend/tests/test_custom_map_router.py
@@ -6,8 +6,8 @@ from fastapi.testclient import TestClient
 
 from src.auth.dependencies import require_approved_user
 from src.campaigns.dependencies import require_campaign_access, require_campaign_admin
-from src.custommaps import router as cm_router
-from src.custommaps import service
+from src.custom_maps import router as cm_router
+from src.custom_maps import service
 from src.database import get_db
 from src.main import app
